# docs/openapi.yaml
openapi: 3.0.2
servers:
  - url: https://api.mainnet.hiro.so/
    description: Mainnet
  - url: https://api.nakamoto.testnet.hiro.so/
    description: Nakamoto Testnet
  - url: https://api.testnet.hiro.so/
    description: Testnet
  - url: http://localhost:3999/
    description: Local
info:
  title: Stacks Blockchain API
  version: 'STACKS_API_VERSION'
  description: |
    Welcome to the API reference overview for the [Stacks Blockchain API](https://docs.hiro.so/stacks-blockchain-api).

    [Download Postman collection](https://hirosystems.github.io/stacks-blockchain-api/collection.json)
...
paths:
...
  /extended/v1/tx/mempool/dropped:
    get:
      summary: Get dropped mempool transactions
      tags:
        - Transactions
      operationId: get_dropped_mempool_transaction_list
      description: |
        Retrieves all recently-broadcast transactions that have been dropped from the mempool.

        Transactions are dropped from the mempool if:
         * they were stale and awaiting garbage collection or,
         * were expensive,  or
         * were replaced with a new fee
      parameters:
        - name: limit
          in: query
          description: max number of mempool transactions to fetch
          required: false
          schema:
            type: integer
            default: 96
            maximum: 200
        - name: offset
          in: query
          description: index of first mempool transaction to fetch
          required: false
          schema:
            type: integer
            example: 42000
      responses:
        200:
          description: List of dropped mempool transactions
          content:
            application/json:
              schema:
                $ref: ./api/transaction/get-mempool-transactions.schema.json
              example:
                $ref: ./api/transaction/get-mempool-transactions.example.json
...
  /extended/v1/tx/{tx_id}/raw:
    parameters:
      - name: tx_id
        in: path
        description: Hash of transaction
        required: true
        schema:
          type: string
          example: "0x0a411719e3bfde95f9e227a2d7f8fac3d6c646b1e6cc186db0e2838a2c6cd9c0"
    get:
      summary: Get Raw Transaction
      tags:
        - Transactions
      operationId: get_raw_transaction_by_id
      description: |
        Retrieves a hex encoded serialized transaction for a given ID
      responses:
        200:
          description: Hex encoded serialized transaction
          content:
            application/json:
              schema:
                $ref: ./api/transaction/get-raw-transaction.schema.json
              example:
                $ref: ./api/transaction/get-raw-transaction.example.json
        404:
          description: Cannot find transaction for given ID
...
  /extended/v1/tx/block/{block_hash}:
    get:
      deprecated: true
      operationId: get_transactions_by_block_hash
      summary: Transactions by block hash
      description: |
        **NOTE:** This endpoint is deprecated in favor of [Get transactions by block](/api/get-transactions-by-block).

        Retrieves a list of all transactions within a block for a given block hash.
      tags:
        - Transactions
      parameters:
        - name: block_hash
          in: path
          description: Hash of block
          required: true
          schema:
            type: string
            example: "0x0a83d82a65460a9e711f85a44616350280040b75317dbe486a923c1131b5ff99"
        - name: limit
          in: query
          description: max number of transactions to fetch
          required: false
          schema:
            type: integer
            example: 10
        - name: offset
          in: query
          description: index of first transaction to fetch
          required: false
          schema:
            type: integer
            example: 42000
      responses:
        200:
          description: List of Transactions
          content:
            application/json:
              schema:
                $ref: ./api/transaction/get-transactions.schema.json
              example:
                $ref: ./api/transaction/get-transactions.example.json
  /extended/v1/tx/block_height/{height}:
    get:
      deprecated: true
      operationId: get_transactions_by_block_height
      summary: Transactions by block height
      description: |
        **NOTE:** This endpoint is deprecated in favor of [Get transactions by block](/api/get-transactions-by-block).

        Retrieves all transactions within a block at a given height
      tags:
        - Transactions
      parameters:
        - name: height
          in: path
          description: Height of block
          required: true
          schema:
            type: integer
            example: 66119
        - name: limit
          in: query
          description: max number of transactions to fetch
          required: false
          schema:
            type: integer
            example: 10
        - name: offset
          in: query
          description: index of first transaction to fetch
          required: false
          schema:
            type: integer
            example: 42000
        - name: unanchored
          in: query
          description: Include transaction data from unanchored (i.e. unconfirmed) microblocks
          required: false
          schema:
            type: boolean
            example: true
            default: false
      responses:
        200:
          description: List of Transactions
          content:
            application/json:
              schema:
                $ref: ./api/transaction/get-transactions.schema.json
              example:
                $ref: ./api/transaction/get-transactions.example.json
...
  /extended/v1/tx/events:
    get:
      summary: Transaction Events
      description: Retrieves the list of events filtered by principal (STX address or Smart Contract ID), transaction id or event types.
        The list of event types is ('smart_contract_log', 'stx_lock', 'stx_asset', 'fungible_token_asset', 'non_fungible_token_asset').
      tags:
        - Transactions
      operationId: get_filtered_events
      parameters:
        - name: tx_id
          in: query
          description: Hash of transaction
          required: false
          example: "0x29e25515652dad41ef675bd0670964e3d537b80ec19cf6ca6f1dd65d5bc642c5"
          schema:
            type: string
        - name: address
          in: query
          description: Stacks address or a Contract identifier
          required: false
          schema:
            type: string
            example: "ST1HB64MAJ1MBV4CQ80GF01DZS4T1DSMX20ADCRA4"
        - name: limit
          in: query
          description: number of items to return
          required: false
          schema:
            type: integer
            example: 100
        - name: offset
          in: query
          description: number of items to skip
          required: false
          schema:
            type: integer
            example: 42000
        - name: type
          in: query
          description: Filter the events on event type
          required: false
          style: form
          explode: true
          schema:
            type: array
            example: stx_lock
            items:
              type: string
              enum: [smart_contract_log, stx_lock, stx_asset, fungible_token_asset, non_fungible_token_asset]
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./api/transaction/get-transaction-events.schema.json
              example:
                $ref: ./api/transaction/get-transaction-events.example.json
...